Add search and filter

+++ src/Employee Management Mini System/Services/EmployeeService.cs
@@ -1,4 +1,4 @@
-﻿using Employee_Management_Mini_System.Models;
+using Employee_Management_Mini_System.Models;
 using Employee_Management_Mini_System.Models.JunctionTables;
 using Employee_Management_Mini_System.Repositories.Contracts;
 using Employee_Management_Mini_System.Services.Contracts;
@@ -20,29 +20,31 @@ namespace Employee_Management_Mini_System.Services
 			_employeeRepository.Create(employee);
 		}
 
-		public async void DeleteEmployee(int employeeId)
+		public async Task DeleteEmployee(int employeeId)
 		{
 			Employee? employee = await _employeeRepository.GetByCondition(e => e.Id == employeeId).SingleOrDefaultAsync();
-			_employeeRepository.Delete(employee);
+			if (employee != null)
+			{
+				_employeeRepository.Delete(employee);
+			}
 		}
 
-		public async void UpdateeEmployee(int employeeId)
+		public void UpdateeEmployee(Employee employee)
 		{
-			Employee? employee = await _employeeRepository.GetByCondition(e => e.Id == employeeId).SingleOrDefaultAsync();
 			_employeeRepository.Update(employee);
 		}
 
 		public async Task<PagedData<Employee>> GetEmployees(int PageNumber,int PageSize = 10)
 		{
 			var totalCount = await _employeeRepository.GetAll().CountAsync();
-			var employees = await _employeeRepository.GetAll().Skip((PageNumber - 1) * PageSize).Take(PageSize).ToListAsync();
+			var employees = await _employeeRepository.GetAll().Include(e => e.Department).Skip((PageNumber - 1) * PageSize).Take(PageSize).ToListAsync();
 			return  new PagedData<Employee> { Data = employees, TotalCount = totalCount };
 		}
 
 		public async Task<PagedData<Employee>> GetEmployeesByCondition(Expression<Func<Employee, bool>> predicate, int PageNumber, int PageSize = 10)
 		{
 			var totalCount = await _employeeRepository.GetByCondition(predicate).CountAsync();
-			var employees = await _employeeRepository.GetByCondition(predicate).Skip((PageNumber - 1) * PageSize).Take(PageSize).ToListAsync();
+			var employees = await _employeeRepository.GetByCondition(predicate).Include(e => e.Department).Skip((PageNumber - 1) * PageSize).Take(PageSize).ToListAsync();
 			return new PagedData<Employee> { Data = employees, TotalCount = totalCount };
 		}
